+++ src/quest.c
@@ -216,7 +216,7 @@ void do_quest( CHAR_DATA *ch, char *argument )
 			sprintf( buf, "&C[&R%5d &P%10d&C] &C[&R%*s&C]\n\r", 
 					quest_table[cnt].level,
 					quest_table[cnt].qp, 
-					52 + strlen( quest_table[cnt].colorname ) - strlen_wo_col( quest_table[cnt].colorname ),
+					(int)(52 + strlen( quest_table[cnt].colorname ) - strlen_wo_col( quest_table[cnt].colorname )),
 					quest_table[cnt].colorname );
 			strcat( result, buf );
 		}
